+++ tests/E2E/UriOverrideTest.php
@@ -20,7 +20,7 @@ class UriOverrideTest extends TestCase
      */
     public function testOverrideWithoutTrailingSlash()
     {
-        $client = new Client('e5e8d4c5c122cf28fc1af3ff77d28bef78a3952957f15067bbc29f2f0dde0b50', 'cfexampleapi', false, null, [
+        $client = new Client('e5e8d4c5c122cf28fc1af3ff77d28bef78a3952957f15067bbc29f2f0dde0b50', 'cfexampleapi', 'master', false, null, [
             'baseUri' => 'https://preview.contentful.com',
         ]);
         $space = $client->getSpace();
@@ -34,7 +34,7 @@ class UriOverrideTest extends TestCase
      */
     public function testOverrideWithTrailingSlash()
     {
-        $client = new Client('e5e8d4c5c122cf28fc1af3ff77d28bef78a3952957f15067bbc29f2f0dde0b50', 'cfexampleapi', false, null, [
+        $client = new Client('e5e8d4c5c122cf28fc1af3ff77d28bef78a3952957f15067bbc29f2f0dde0b50', 'cfexampleapi', 'master', false, null, [
             'baseUri' => 'https://preview.contentful.com/',
         ]);
         $space = $client->getSpace();