+++ tests/Unit/ClientTest.php
@@ -33,7 +33,7 @@ class ClientTest extends TestCase
 
     public function testIsPreview()
     {
-        $client = new Client('b4c0n73n7fu1', 'cfexampleapi', true);
+        $client = new Client('b4c0n73n7fu1', 'cfexampleapi', 'master', true);
 
         $this->assertTrue($client->isPreview());
         $this->assertSame('PREVIEW', $client->getApi());
@@ -57,7 +57,7 @@ class ClientTest extends TestCase
      */
     public function testInvalidCachePool()
     {
-        new Client('b4c0n73n7fu1', 'cfexampleapi', false, null, [
+        new Client('b4c0n73n7fu1', 'cfexampleapi', 'master', false, null, [
             'cache' => new \stdClass(),
         ]);
     }